refactor(repo): split multi-primary export files in packages/plugins
