+++ packages/components/src/toast/index.ts
@@ -1,2 +1,3 @@
-export { ToastProvider, useToast } from './ToastProvider.js';
+export { ToastProvider } from './ToastProvider.js';
 export type { ToastConfig } from './ToastProvider.js';
+export { useToast, useToastNotifications } from './useToast.js';
