+++ packages/plugin-ui-sdk/src/hooks/useTreeNodeUpdater.ts
@@ -93,31 +93,6 @@ const stableStringify = (value: unknown): string => {
   }
 };
 
-export const createTreeNodeUpdaterActions = <
-  TPayload extends PeerEntity<TreeNodeData> = PeerEntity<TreeNodeData>
->(
-  updateDraft: (data: Partial<TreeNodeUpdaterState<TPayload>>) => void
-) => {
-  const updatePayload = (patch: Partial<TPayload>, base?: TPayload) => {
-    const next = { ...(base ?? ({} as TPayload)), ...patch } as TPayload;
-    updateDraft({ draftData: next });
-  };
-  const updateMetadata = (patch: Partial<TreeNodeMetadata>, base?: TreeNodeMetadata) => {
-    const fallback: TreeNodeMetadata = { name: '', description: '', tags: [] };
-    updateDraft({ draftMetadata: { ...(base ?? fallback), ...patch } });
-  };
-  const updatePayloadAndMetadata = (
-    payloadPatch: Partial<TPayload>,
-    metadataPatch: Partial<TreeNodeMetadata>,
-    base?: { payload?: TPayload; metadata?: TreeNodeMetadata }
-  ) => {
-    updatePayload(payloadPatch, base?.payload);
-    updateMetadata(metadataPatch, base?.metadata);
-  };
-
-  return { updatePayload, updateMetadata, updatePayloadAndMetadata };
-};
-
 export function useTreeNodeUpdater<
   TPayload extends PeerEntity<TreeNodeData> = PeerEntity<TreeNodeData>
 >({
