+++ packages/ui/tabular-extract/src/hooks/useTabularFilter.ts
@@ -5,7 +5,7 @@
 
 import { useCallback, useEffect, useState } from 'react';
 import { useTabularApi } from '../context/TabularContext';
-import type { TabularDataResult, TabularFilterRule, TabularSelectionConfig } from '../types/index';
+import type { TabularDataResult, TabularFilterRule } from '../types/index';
 
 /**
  * Options for useTabularFilter hook
@@ -282,174 +282,5 @@ export const useTabularFilter = (options: UseTabularFilterOptions): UseTabularFi
     toggleRule,
     refreshPreview,
     setRules,
-  };
-};
-
-/**
- * Options for useCSVSelection hook
- */
-export interface UseTabularSelectionOptions {
-  /** Table metadata ID */
-  tableMetadataId: string;
-  /** Filter rules to apply */
-  filterRules: TabularFilterRule[];
-  /** Initial selection configuration */
-  initialSelection?: Partial<TabularSelectionConfig>;
-  /** Number of preview rows */
-  previewRowCount?: number;
-}
-
-/**
- * Result of useCSVSelection hook
- */
-export interface UseTabularSelectionResult {
-  /** Current selection configuration */
-  selection: TabularSelectionConfig;
-  /** Preview data with current selection */
-  previewData: TabularDataResult | null;
-  /** Loading atoms */
-  loading: boolean;
-  /** Error message */
-  error: string | null;
-
-  /** Set key column */
-  setKeyColumn: (column: string | undefined) => void;
-  /** Set value columns */
-  setValueColumns: (columns: string[]) => void;
-  /** Add value column */
-  addValueColumn: (column: string) => void;
-  /** Remove value column */
-  removeValueColumn: (column: string) => void;
-  /** Set custom mappings */
-  setCustomMappings: (mappings: TabularSelectionConfig['customMappings']) => void;
-  /** Update entire selection */
-  updateSelection: (updates: Partial<TabularSelectionConfig>) => void;
-  /** Refresh preview */
-  refreshPreview: () => Promise<void>;
-}
-
-/**
- * Hook for managing CSV column selection
- */
-export const useTabularSelection = (options: UseTabularSelectionOptions): UseTabularSelectionResult => {
-  const { tableMetadataId, filterRules, initialSelection = {}, previewRowCount = 10 } = options;
-
-  const tabularApi = useTabularApi();
-
-  const [selection, setSelection] = useState<TabularSelectionConfig>({
-    keyColumn: undefined,
-    valueColumns: [],
-    filterRules,
-    customMappings: [],
-    ...initialSelection,
-  });
-
-  const [previewData, setPreviewData] = useState<TabularDataResult | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  /**
-   * Fetch preview with current selection
-   */
-  const fetchPreview = useCallback(async () => {
-    if (!tableMetadataId) return;
-
-    try {
-      setLoading(true);
-      setError(null);
-
-      const data = await tabularApi.getFilteredPreview(
-        tableMetadataId,
-        filterRules.filter((rule) => rule.enabled !== false),
-        previewRowCount,
-      );
-
-      setPreviewData(data);
-    } catch (err) {
-      const message = err instanceof Error ? err.message : 'Failed to fetch preview';
-      setError(message);
-      setPreviewData(null);
-    } finally {
-      setLoading(false);
-    }
-  }, [tabularApi, tableMetadataId, filterRules, previewRowCount]);
-
-  /**
-   * Set key column
-   */
-  const setKeyColumn = useCallback((column: string | undefined) => {
-    setSelection((prev) => ({ ...prev, keyColumn: column }));
-  }, []);
-
-  /**
-   * Set value columns
-   */
-  const setValueColumns = useCallback((columns: string[]) => {
-    setSelection((prev) => ({ ...prev, valueColumns: columns }));
-  }, []);
-
-  /**
-   * Add value column
-   */
-  const addValueColumn = useCallback((column: string) => {
-    setSelection((prev) => ({
-      ...prev,
-      valueColumns: [...prev.valueColumns, column],
-    }));
-  }, []);
-
-  /**
-   * Remove value column
-   */
-  const removeValueColumn = useCallback((column: string) => {
-    setSelection((prev) => ({
-      ...prev,
-      valueColumns: prev.valueColumns.filter((col) => col !== column),
-    }));
-  }, []);
-
-  /**
-   * Set custom mappings
-   */
-  const setCustomMappings = useCallback((mappings: TabularSelectionConfig['customMappings']) => {
-    setSelection((prev) => ({ ...prev, customMappings: mappings }));
-  }, []);
-
-  /**
-   * Update entire selection
-   */
-  const updateSelection = useCallback((updates: Partial<TabularSelectionConfig>) => {
-    setSelection((prev) => ({ ...prev, ...updates }));
-  }, []);
-
-  /**
-   * Refresh preview
-   */
-  const refreshPreview = useCallback(async () => {
-    await fetchPreview();
-  }, [fetchPreview]);
-
-  // Update filter rules in selection when they change
-  useEffect(() => {
-    setSelection((prev) => ({ ...prev, filterRules }));
-  }, [filterRules]);
-
-  // Auto-refresh preview when selection changes
-  useEffect(() => {
-    fetchPreview();
-  }, [fetchPreview]);
-
-  return {
-    selection,
-    previewData,
-    loading,
-    error,
-    setKeyColumn,
-    setValueColumns,
-    addValueColumn,
-    removeValueColumn,
-    setCustomMappings,
-    updateSelection,
-    refreshPreview,
   };
 };
