+++ plugins/location-plugin/src/common/i18n/index.ts
@@ -1,4 +1,5 @@
 import { i18n as globalI18n } from '@hierarchidb/ui-i18n';
+export { formatBytes, formatNumber } from './formatters.js';
 import en from '~/ui/locales/en.json' with { type: 'json' };
 import ja from '~/ui/locales/ja.json' with { type: 'json' };
 
@@ -433,21 +434,4 @@ export const useTranslation = (ns: string = 'location-plugin') => {
   const t = (key: string, fallback?: string) =>
     String(globalI18n.t(key, { ns, defaultValue: fallback ?? key }));
   return { t, translations, locale };
-};
-
-export const formatBytes = (bytes: number, locale: SupportedLocale = detectLocale()): string => {
-  if (!Number.isFinite(bytes)) return '0 B';
-  const units = ['B', 'KB', 'MB', 'GB', 'TB'];
-  let value = bytes;
-  let unitIndex = 0;
-  while (value >= 1024 && unitIndex < units.length - 1) {
-    value /= 1024;
-    unitIndex += 1;
-  }
-  const formatter = new Intl.NumberFormat(locale, { maximumFractionDigits: 2 });
-  return `${formatter.format(value)} ${units[unitIndex]}`;
-};
-
-export const formatNumber = (value: number, locale: SupportedLocale = detectLocale()): string => {
-  return new Intl.NumberFormat(locale).format(value);
 };
